Small refactor of the Consumer stream impl
It was previously impossible to execute the `consume` integration test
as is, because you had to be able to poll the consumer stream while
polling for the ack / reject future, which isn't possible when only
using `Runtime::block_on`. We're now using the channel directly from
`batch_rabbitmq::Delivery`, though that did reveal a new bug that broke
the worker example.